Move reader stuff from utils to reader package (#277)
