+++ kernel/src/main/scala/me/rexim/morganey/util/package.scala
@@ -4,47 +4,13 @@ import java.io.{File, FileInputStream, InputStreamReader, Reader}
 import java.net.URL
 import java.nio.charset.StandardCharsets.UTF_8
 import java.util.regex.Pattern
-
-import me.rexim.morganey.syntax.{LambdaParser, LambdaParserException}
-
 import scala.util._
 
 package object util {
-
   // TODO(#276): move validRegex to Commands
   def validRegex(regex: String): Option[String => Boolean] =
     Try {
       val pattern = Pattern.compile(regex)
       s: String => pattern.matcher(s).matches()
     }.toOption
-
-
-  // TODO(#277): move reader stuff to me.rexim.morganey.reader
-  def reader(path: String): Try[Reader] = reader(new File(path))
-
-  def reader(file: File): Try[Reader] = {
-    val inputStream = Try(new FileInputStream(file))
-    inputStream.map(new InputStreamReader(_, UTF_8))
-  }
-
-  def reader(url: URL): Try[Reader] = {
-    Try(new InputStreamReader(url.openStream(), UTF_8))
-  }
-
-  def withReader[T](path: String)(f: Reader => Try[T]): Try[T] =
-    withReader(new File(path))(f)
-
-  def withReader[T](url: URL)(f: Reader => Try[T]): Try[T] =
-    reader(url).flatMap { reader =>
-      val result = f(reader)
-      reader.close()
-      result
-    }
-
-  def withReader[T](file: File)(f: Reader => Try[T]): Try[T] =
-    reader(file).flatMap { reader =>
-      val result = f(reader)
-      reader.close()
-      result
-    }
 }
